@@ -3,27 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "apexcharts": "^3.19.0",
-    "axios": "^0.19.2",
-    "bootstrap": "^4.4.1",
-    "country-emoji": "^1.5.3",
+    "@reduxjs/toolkit": "^2.2.7",
+    "@testing-library/jest-dom": "^6.4.8",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
+    "apexcharts": "^3.52.0",
+    "axios": "^1.7.4",
+    "bootstrap": "^5.3.3",
+    "country-emoji": "^1.5.6",
     "font-awesome": "^4.7.0",
-    "mdbreact": "^4.26.1",
-    "react": "^16.13.1",
-    "react-apexcharts": "^1.3.7",
-    "react-data-components": "^1.2.0",
-    "react-dom": "^16.13.1",
+    "react": "^18.3.1",
+    "react-apexcharts": "^1.3.9",
+    "react-dom": "^18.3.1",
     "react-jvectormap": "0.0.16",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1",
-    "reactstrap": "^8.4.1",
-    "recharts": "^1.8.5",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.26.1",
+    "react-scripts": "^5.0.1",
+    "react-table": "^7.8.0",
+    "reactstrap": "^9.2.2",
+    "recharts": "^2.12.7",
+    "redux": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
